Fix mark as complete action
